navigation between notificationDaily and NotificationHistory
