+++ safe1/src/views/index.tsx
@@ -1,5 +1,6 @@
 import Home from './home'
 import Dashboard from './dashboard'
 import Login from './login'
-import Notification from './notification'
-export {Home, Dashboard, Notification , Login}
+import NotificationHistory from './NotificationHistory'
+import NotificationDaily from './NotificationDaily'
+export {Home, Dashboard, NotificationHistory, NotificationDaily , Login}
